Add files via upload
Press R click to place blocks or L click to destroy blocks (this is available by pressing the button 1); Press 2 for pallete (you can press esc, choose a color, then press L click to color blocks or just press 1 and put color blocks) Have fun! You have Save World and import world. If u go in a spinning rotation bug, it goes continiously is the W or A or S or D dirrection press the opposite key to stop, then unselect or select the direction that is moving you to, and thats it. The coordonates are glitched, I couldn't solve this problem, if someone knows, a little help is usefull